requirements.txt
openai
python-dotenv
tree-sitter